+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes
+- Fixed the key suggestions in the filter based on [this issue](https://github.com/maplibre/maputnik/issues/2091)
 - _...Add new stuff here..._
 
 ## 3.1.0